requirements.txt
dlib
einops
numpy
fastai
google-api-python-client
google-auth-oauthlib
levenshtein
matplotlib
omegaconf
pandas
Pillow
scipy
schedulefree
timm>=0.9
wandb
torchinfo
pyopengl
glfw
opencv-python
# torch==2.1
# openmim
